requirements.txt
python-dotenv
twitchio
discord.py